% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID_urine}
\alias{COVID_urine}
\title{Metabolite signatures of COVID obtained from Human urine}
\format{
A list of data frames
}
\usage{
data(COVID_urine)
}
\value{
A list object
}
\description{
A list containing metabolite signatures
}
\references{
Bi, Xiaojie, et al.
"Proteomic and metabolomic profiling of urine uncovers immune responses in patients with COVID-19."
Cell reports 38.3 (2022).
}
\keyword{datasets}
