% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature2gene.R
\name{.extract_sig_meta}
\alias{.extract_sig_meta}
\title{Title Subset the GEM metabolite data frame}
\usage{
.extract_sig_meta(signature, GEM_meta_df, reference_key = "refmet_name")
}
\arguments{
\item{signature}{a character vector contains standardized names of signatures}

\item{GEM_meta_df}{the GEM metabolite data frame}

\item{reference_key}{the key which is used to map the signature and metabolites in the GEM}
}
\value{
a subset of the GEM metabolite data frame
}
\description{
Title Subset the GEM metabolite data frame
}
\keyword{internal}
