% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature2gene.R
\name{.extract_standardized_name}
\alias{.extract_standardized_name}
\title{Title Extract the standardized names from the signature data frames}
\usage{
.extract_standardized_name(signature_df, reference_key = "refmet_name")
}
\arguments{
\item{signature_df}{signature data frame}

\item{reference_key}{the key which is used to map the signature and metabolites in the GEM}
}
\value{
a character vector contains standardized names of signatures
}
\description{
Title Extract the standardized names from the signature data frames
}
\keyword{internal}
