% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature2gene.R
\name{.gene_association}
\alias{.gene_association}
\title{Title Gene association of each signature}
\usage{
.gene_association(
  mapped_metabolite_signature,
  m2g,
  reference_key = "refmet_name"
)
}
\arguments{
\item{mapped_metabolite_signature}{one element in "mapped_metabolite_signatures"}

\item{m2g}{association list}

\item{reference_key}{the key which is used to map the signature and metabolites in the GEM}
}
\value{
a updated 'mapped_metabolite_signature' with "gene_association" column
}
\description{
Title Gene association of each signature
}
\keyword{internal}
