% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.hyper_enrichment}
\alias{.hyper_enrichment}
\title{Hypergeometric test}
\usage{
.hyper_enrichment(
  gene_table,
  genesets,
  genesets_name = "unknown",
  min_metabolite,
  background = 1234567
)
}
\arguments{
\item{gene_table}{the gene table from hypeR-GEM object}

\item{genesets}{a list of genesets}

\item{genesets_name}{name of the geneset,e.g "KEGG"}

\item{min_metabolite}{minimum number/ratio of metabolite that drives the enrichment}

\item{background}{background parameter of hypergeometric test}
}
\value{
a list
}
\description{
Hypergeometric test
}
\keyword{internal}
