% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{.multi_dots_plot}
\alias{.multi_dots_plot}
\title{Create a multi dots plot}
\usage{
.multi_dots_plot(
  multihyp_data,
  top = 20,
  abrv = 50,
  size_by = c("genesets", "significance", "none"),
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  title = ""
)
}
\arguments{
\item{multihyp_data}{data in hypeR_GEM_enrichments}

\item{top}{Limit number of genesets shown}

\item{abrv}{Abbreviation length of genesetlabels}

\item{size_by}{Size dots by e.g. c("genesets", "significance", "none")}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value e.g. c("fdr", "pval")}

\item{title}{title Plot title}
}
\description{
Create a multi dots plot
}
\keyword{internal}
