% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{.rctbl}
\alias{.rctbl}
\title{Title Reactable table for hypeR_GEM_enrichment object}
\usage{
.rctbl(hypeR_GEM_enrichment, type = c("inner", "outer"), fdr_cutoff = 0.05)
}
\arguments{
\item{hypeR_GEM_enrichment}{A hypeR_GEM_enrichment from a single signature}

\item{type}{Use style class for outer or inner tables}
}
\value{
a reactable
}
\description{
Title Reactable table for hypeR_GEM_enrichment object
}
\keyword{internal}
