% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EL_signature}
\alias{EL_signature}
\title{A template of metabolite signatures}
\format{
A list of dataframes
}
\usage{
EL_signature
}
\description{
A list containing metabolite signatures
}
\keyword{datasets}
