% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
enrichment(
  hypeR_GEM_obj,
  genesets,
  genesets_name = "unknown",
  method = c("unweighted", "weighted"),
  weighted_by = "one_minus_fdr",
  a = -1,
  b = -1,
  sigmoid_transformation = TRUE,
  min_metabolite = 0,
  background = 1234567
)
}
\arguments{
\item{hypeR_GEM_obj}{an hypeR_GEM object, the output of "signature2gene" function}

\item{genesets}{a list of genesets}

\item{genesets_name}{name of the geneset,e.g "KEGG"}

\item{method}{enrichment method}

\item{weighted_by}{the column name in the gene_table of hypeR_GEM_obj that represents the weight of each gene}

\item{a}{proportional to the smoothness of the sigmoid function, default a = -1}

\item{b}{the half-point threshold of the sigmoid function, default b = -1 (p-value = 0.1 as half-point)}

\item{sigmoid_transformation}{logical; when method == "weighted", if TRUE apply .sigmoid_transformation() to `weighted_by`; if FALSE use the raw values}

\item{min_metabolite}{minimum number of metabolite that drives the enrichment of a pathway}

\item{background}{background parameter of hypergeometric test}
}
\value{
a list of type of statistical test and data
}
\description{
Enrichment analysis
}
