% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{enrichment_plot}
\alias{enrichment_plot}
\title{Title Enrichment analysis visualization}
\usage{
enrichment_plot(
  hypeR_GEM_enrichment,
  top = 20,
  abrv = 50,
  size_by = c("genesets", "significance", "none"),
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  title = ""
)
}
\arguments{
\item{hypeR_GEM_enrichment}{an hypeR_GEM object, the output of "signature2gene" function}

\item{title}{}
}
\value{
a ggplot object
}
\description{
Title Enrichment analysis visualization
}
