% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Title extract tables from a given genome scale metabolic model(GEM) of ".mat" format}
\usage{
extract_tables(con)
}
\arguments{
\item{con}{a connection or the path of the GEM}
}
\value{
a list of tables
}
\description{
Title extract tables from a given genome scale metabolic model(GEM) of ".mat" format
}
\examples{
con = file.path("data/Recon3D.mat")
tables = extract_tables(con)
}
