% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsets_list}
\alias{gsets_list}
\title{A list of enriched gene-sets in HALLMARK,KEGG, and REACTOME derived from "hypeR_object_NECS"}
\format{
a list object
}
\usage{
data(gsets_list)
}
\value{
a list object
}
\description{
A a list object
}
\references{
Federico, Anthony, and Stefano Monti.
"hypeR: an R package for geneset enrichment workflows."
Bioinformatics 36.4 (2020): 1307-1308.
}
\keyword{datasets}
