% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Human_GEM_tables_directed}
\alias{Human_GEM_tables_directed}
\title{Directed mapping tables from Human GEM}
\format{
A nested list of dataframes and lists
}
\usage{
Human_GEM_tables_directed
}
\description{
A list of tables of genes, metabolites, and reactions
}
\keyword{datasets}
