% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Human_ig}
\alias{Human_ig}
\title{Network of Human GEM}
\format{
An object of class \code{igraph} from \code{igraph} package
}
\usage{
data(Human_ig)
}
\value{
An igraph object
}
\description{
An igraph object consists of a directed network
}
\keyword{datasets}
