% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypeR_object_NECS}
\alias{hypeR_object_NECS}
\title{hypeR-GEM object of age-associated metabolites obtained from New England Centenarian Study (NECS)}
\format{
a list object
}
\usage{
data(hypeR_object_NECS)
}
\value{
a list object
}
\description{
A a list object
}
\references{
TBD
}
\keyword{datasets}
