% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypGEM_to_excel.R
\name{hypGEM_to_excel}
\alias{hypGEM_to_excel}
\title{Writes the enrichment results out to an Excel workbook}
\usage{
hypGEM_to_excel(
  hypGEM_obj,
  file_path,
  cols = NULL,
  fdr_cutoff = 1,
  do_sort = FALSE,
  versioning = TRUE
)
}
\arguments{
\item{hypGEM_obj}{A hypeR-GEM object from enrichment analysis}

\item{file_path}{the path to save the output file}

\item{cols}{columns}

\item{fdr_cutoff}{fdr threshold}

\item{do_sort}{logical parameter}

\item{versioning}{logical parameter}
}
\value{
a excel file
}
\description{
Writes the enrichment results out to an Excel workbook
}
