% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolon_msets_list}
\alias{metabolon_msets_list}
\title{A list of enriched Metabolon-based metabolite sets (MSets) obtained from age-associated metabolites in NECS}
\format{
a list object
}
\usage{
data(metabolon_msets_list)
}
\value{
a list object
}
\description{
A a list object
}
\references{
Evans, A. M., DeHaven, C. D., Barrett, T., Mitchell, M. & Milgram, E.
Integrated, Nontargeted Ultrahigh Performance Liquid Chromatography/Electrospray Ionization
Tandem Mass Spectrometry Platform for the Identification and Relative Quantification of
the Small-Molecule Complement of Biological Systems. Anal. Chem. 81, 6656–6667 (2009).
}
\keyword{datasets}
