% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rctbls}
\alias{rctbls}
\title{Title Reactable table for hypeR_GEM_enrichment object}
\usage{
rctbls(hypeR_GEM_enrichments)
}
\arguments{
\item{hypeR_GEM_enrichments}{A hypeR_GEM_enrichments from multiple signatures}
}
\value{
a nested reactable
}
\description{
Title Reactable table for hypeR_GEM_enrichment object
}
