% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reactome}
\alias{Reactome}
\title{Reactome genesets}
\format{
An object of class \code{hypeR} from \code{hypeR} package
}
\usage{
data(Reactome)
}
\value{
An hypeR object
}
\description{
A hypeR object
}
\references{
Federico, Anthony, and Stefano Monti.
"hypeR: an R package for geneset enrichment workflows."
Bioinformatics 36.4 (2020): 1307-1308.
}
\keyword{datasets}
