% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNet.R
\name{visNet}
\alias{visNet}
\title{Title Query and plot the sub-graph}
\usage{
visNet(
  ig,
  metabolite_query = NULL,
  gene_query = NULL,
  metabolite_by = c("symbol", "refmet_name", "HMDB_ID", "name"),
  gene_by = c("symbol", "ensemble_ID", "name"),
  focus = c("all", "metabolite", "pathway"),
  edge_label = NULL,
  main = "A simple reaction-based network",
  metabolite_query_color = "#FF9999",
  gene_query_color = "#56B4E9",
  associated_metabolite_color = "#808080",
  associated_gene_color = "#808080",
  metabolite_shape = "dot",
  gene_shape = "diamond",
  hover = TRUE,
  stabilization = FALSE,
  igraph_layout = "layout_with_fr",
  layout_seed = 42
)
}
\arguments{
\item{ig}{the igraph object derived from a given GEM}

\item{metabolite_query}{a character vector contains metabolites which are consistent with V(ig)$symbol or V(ig)$name, for metabolite, the "symbol" should be consistent with "fullname" column in hypeR_GEM_obj$mapped_metabolite_signatures}

\item{gene_query}{a character vector contains enzyme-coding genes which are consistent with V(ig)$symbol or V(ig)$name}

\item{metabolite_by}{type of symbols used by metabolite_query}

\item{gene_by}{type of symbols used by gene_query}

\item{focus}{focus of the subgraph query}

\item{edge_label}{An edge attribute in E(ig) that can be used to represent edge labels}

\item{main}{the title of the network}

\item{metabolite_query_color}{the node color of "metabolite_query"}

\item{gene_query_color}{the node color of "gene_query"}

\item{associated_metabolite_color}{the node color of metabolites other than "metabolite_query"}

\item{associated_gene_color}{the node color of genes other than "gene_query"}

\item{metabolite_shape}{the node shape of metabolites}

\item{gene_shape}{the node shape of genes}

\item{hover}{logical variable}

\item{stabilization}{logical variable}

\item{igraph_layout}{the layout algorithm available in "igraph", see https://igraph.org/r/html/1.3.0/layout_.html}

\item{layout_seed}{random seed of the network layout}
}
\value{
a list of: ig = node-induced sub-graph, p = a visNetwork object
}
\description{
Title Query and plot the sub-graph
}
