% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Worm_GEM_tables}
\alias{Worm_GEM_tables}
\title{Undirected mapping tables from Worm GEM}
\format{
A nested list of dataframes and lists
}
\usage{
Worm_GEM_tables
}
\description{
A list of tables of genes, metabolites, and reactions
}
\keyword{datasets}
