% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Worm_GEM_tables_di}
\alias{Worm_GEM_tables_di}
\title{Directional mapping tables of Worm GEM}
\format{
A list of data frames and lists
}
\usage{
data(Worm_GEM_tables_di)
}
\value{
A list object
}
\description{
A list of tables of genes, metabolites, and reactions
}
\references{
Wang, Hao, et al.
"Genome-scale metabolic network reconstruction of model animals as a platform for translational research."
Proceedings of the National Academy of Sciences 118.30 (2021): e2102344118.
}
\keyword{datasets}
