% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zebrafish_GEM_tables}
\alias{Zebrafish_GEM_tables}
\title{Undirectional mapping tables of Zebrafish GEM}
\format{
A list of data frames and lists
}
\usage{
data(Zebrafish_GEM_tables)
}
\value{
A list object
}
\description{
A list of tables of genes, metabolites, and reactions
}
\references{
Wang, Hao, et al.
"Genome-scale metabolic network reconstruction of model animals as a platform for translational research."
Proceedings of the National Academy of Sciences 118.30 (2021): e2102344118.
}
\keyword{datasets}
