% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.hyper_enrichment}
\alias{.hyper_enrichment}
\title{Title Hypergeometric test}
\usage{
.hyper_enrichment(
  gene_table,
  genesets,
  filter_threshold,
  genesets_name = "unknown",
  background = 1234567
)
}
\arguments{
\item{genesets}{a list of genesets}

\item{background}{background parameter of hypergeometric test}

\item{unweighted_signature}{a character vector of signature}
}
\value{
a list
}
\description{
Title Hypergeometric test
}
\keyword{internal}
