% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.hyper_enrichment}
\alias{.hyper_enrichment}
\title{Title Hypergeometric test}
\usage{
.hyper_enrichment(unweighted_signature, genesets, background = 1234567)
}
\arguments{
\item{unweighted_signature}{a character vector of signature}

\item{genesets}{a list of genesets}

\item{background}{background parameter of hypergeometric test}
}
\value{
a list
}
\description{
Title Hypergeometric test
}
\keyword{internal}
