% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature2gene.R
\name{.meta2gene}
\alias{.meta2gene}
\title{Title Map metabolic signature to ranked genes}
\usage{
.meta2gene(signature, GEM_tables, background, reference_key = "refmet_name")
}
\value{
a data frame
}
\description{
Title Map metabolic signature to ranked genes
}
\keyword{internal}
