% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{.multi_dots_plot}
\alias{.multi_dots_plot}
\title{Title Create a multi dots plot}
\usage{
.multi_dots_plot(
  multihyp_data,
  top = 20,
  abrv = 50,
  size_by = c("genesets", "significance", "none"),
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  title = ""
)
}
\description{
Title Create a multi dots plot
}
\keyword{internal}
