% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.weighted_hyper_enrichment}
\alias{.weighted_hyper_enrichment}
\title{Weighted hypergeometric test}
\usage{
.weighted_hyper_enrichment(
  gene_table,
  genesets,
  genesets_name = "unknown",
  min_metabolite,
  weights = "one_minus_fdr",
  background = 1234567
)
}
\arguments{
\item{gene_table}{the gene table from hypeR-GEM object}

\item{genesets}{a list of genesets}

\item{genesets_name}{name of the geneset,e.g "KEGG"}

\item{min_metabolite}{minimum number/ratio of metabolite that drives the enrichment}

\item{background}{background parameter of hypergeometric test}
}
\value{
a list
}
\description{
Weighted hypergeometric test
}
\keyword{internal}
