% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.weighted_hyper_enrichment}
\alias{.weighted_hyper_enrichment}
\title{Title Weighted hypergeometric test}
\usage{
.weighted_hyper_enrichment(
  gene_table,
  genesets,
  filter_threshold,
  genesets_name = "unknown",
  weights = "one_minus_fdr",
  background = 1234567
)
}
\arguments{
\item{genesets}{a list of genesets}

\item{background}{background parameter of hypergeometric test}

\item{weighted_signature}{named vector, name = gene symbol, value = weight and should be between 0 and 1}
}
\value{
a list
}
\description{
Title Weighted hypergeometric test
}
\keyword{internal}
