% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Title Enrichment analysis}
\usage{
enrichment(
  hypeR_GEM_obj,
  genesets,
  genesets_name = "unknown",
  method = c("unweighted", "weighted", "kstest"),
  weights = "one_minus_fdr",
  metabolite_hits_filter_threshold = 0,
  background = 1234567
)
}
\arguments{
\item{hypeR_GEM_obj}{an hypeR_GEM object, the output of "signature2gene" function}

\item{genesets}{a list of genesets}

\item{method}{enrichment method}

\item{weights}{the column name in the gene_table of hypeR_GEM_obj that represents the weight of each gene}

\item{background}{background parameter of hypergeometric test}

\item{metetabolite_hits_filter_threshold}{filter out enriched pathways that are driven by a small number/ratio of metabolites, if 0 < threshold < 1, we filter by ratio, else, we filter by number. If threshold = 0, no filter}
}
\value{
a list of type of statistical test and data
}
\description{
Title Enrichment analysis
}
