% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{enrichment_plot}
\alias{enrichment_plot}
\title{Title Enrichment analysis visualization}
\usage{
enrichment_plot(
  hypeR_GEM_enrichments,
  top = 20,
  abrv = 50,
  size_by = c("genesets", "significance", "none"),
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  title = ""
)
}
\arguments{
\item{top}{Limit number of genesets shown}

\item{abrv}{Abbreviation length of geneset labels}

\item{size_by}{Size dots by e.g. c("genesets", "significance", "none")}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value e.g. c("fdr", "pval")}

\item{title}{Plot title}

\item{hypeR_GEM_enrichment}{A hypeR_GEM_enrichments from single/multiple signatures, the output of "hypeR.GEM::enrichment()"}
}
\value{
a ggplot object
}
\description{
Title Enrichment analysis visualization
}
