% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Title extract tables from a given genome scale metabolic model(GEM) of ".mat" format}
\usage{
extract_tables(con, directional = FALSE)
}
\arguments{
\item{con}{a connection or the path of the GEM, which should be in .mat format}

\item{directional}{logical parameter, if TRUE, map metabolites to reactions where these metabolites are product only}
}
\value{
a list of tables
}
\description{
Title extract tables from a given genome scale metabolic model(GEM) of ".mat" format
}
