% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypeR_object_NECS}
\alias{hypeR_object_NECS}
\title{Example hypeR-GEM object (NECS)}
\format{
A list with components required by hypeR-GEM downstream functions.
}
\usage{
data(hypeR_object_NECS)
}
\description{
A small precomputed hypeR-GEM result derived from age-associated metabolites
in the New England Centenarian Study (NECS), used for examples and tests.
}
\references{
TBD
}
\keyword{datasets}
