% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypGEM_to_excel.R
\name{hypGEM_filter}
\alias{hypGEM_filter}
\title{hypGEM_filter}
\usage{
hypGEM_filter(hypGEM_obj, cols = NULL, fdr_cutoff = 1, do_sort = FALSE)
}
\description{
Filter the enrichment results by fdr value, and optionally sub-selects result columns
}
