% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypGEM_to_excel.R
\name{hypGEM_to_excel}
\alias{hypGEM_to_excel}
\title{hypGEM_to_excel}
\usage{
hypGEM_to_excel(
  hypGEM_obj,
  file_path,
  cols = NULL,
  fdr_cutoff = 1,
  do_sort = FALSE,
  versioning = TRUE
)
}
\description{
writes the enrichment results out to an Excel workbook
}
