% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolon_msets_list}
\alias{metabolon_msets_list}
\title{Metabolon-based metabolite set list (NECS-derived)}
\format{
A named list of character vectors; each vector contains RefMet (or study) IDs.
}
\usage{
data(metabolon_msets_list)
}
\description{
A curated list of enriched Metabolon metabolite sets (MSets) obtained from
age-associated metabolites in NECS.
}
\references{
Evans AM, DeHaven CD, Barrett T, Mitchell M, Milgram E.
Integrated, Nontargeted Ultrahigh Performance Liquid Chromatography/Electrospray Ionization
Tandem Mass Spectrometry Platform for the Identification and Relative Quantification of
the Small-Molecule Complement of Biological Systems. \emph{Anal Chem}. 2009;81:6656–6667.
}
\keyword{datasets}
