% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rctbls}
\alias{rctbls}
\title{Title Reactable table for hypeR_GEM_enrichment object}
\usage{
rctbls(hypeR_GEM_obj, fdr_cutoff = 0.05)
}
\arguments{
\item{hypeR_GEM_obj}{A list of enrichment results from single/multiple signatures, output of "hypeR.GEM::enrichment()"}

\item{fdr_cutoff}{fdr threshold for geneset enrichment}
}
\value{
a nested reactable
}
\description{
Title Reactable table for hypeR_GEM_enrichment object
}
