% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome}
\alias{reactome}
\title{Reactome genesets}
\format{
a list object
}
\usage{
data(reactome)
}
\value{
a list object
}
\description{
A a list object
}
\references{
Federico, Anthony, and Stefano Monti.
"hypeR: an R package for geneset enrichment workflows."
Bioinformatics 36.4 (2020): 1307-1308.
}
\keyword{datasets}
