% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature2gene.R
\name{signature2gene}
\alias{signature2gene}
\title{Title Map metabolic signatures to enzyme-coding genes}
\usage{
signature2gene(
  signatures,
  species = c("Human", "Mouse", "Rat", "Zebrafish", "Worm", "Other"),
  tables = NULL,
  merge = TRUE,
  reference_key = "refmet_name",
  promiscuous_threshold = NULL,
  ensemble_id = TRUE,
  background = NULL
)
}
\arguments{
\item{signatures}{a list of metabolic signatures, each element is a data frame which has to contain a column with the same name as "reference_key"}

\item{species}{the species of GEM model, if species == 'Other', a user-define GEM_table must be provided}

\item{tables}{species == 'Other', a user-define GEM_table must be provided}

\item{merge}{merged metabolites in different compartment, for example, MAM001_c, MAM001_e, ... -> MAM001}

\item{reference_key}{the column name in the reference table which represents the standardized names (e.g. "refmet_name")}

\item{promiscuous_threshold}{gene association threshold of promiscuous metabolite}

\item{ensemble_id}{if the genes in GEM_tables$gene_df is given by Ensemble, then ensemble_id = TRUE, otherwise, ensemble_id = FALSE}

\item{background}{the background of the gene-specific hypergeometric test, default = NULL = number of merged metabolites}
}
\value{
a list containing two element: "mapped_metabolite_signatures" and "gene_tables" for each signature
}
\description{
Title Map metabolic signatures to enzyme-coding genes
}
